fixed beautiful 3d sphere

+++ srcs/vector_color.c
@@ -6,7 +6,7 @@
 /*   By: lmartin <lmartin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/01 06:45:02 by lmartin           #+#    #+#             */
-/*   Updated: 2019/11/01 07:01:07 by lmartin          ###   ########.fr       */
+/*   Updated: 2019/11/01 13:07:58 by lmartin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,10 +18,10 @@ s_vector	*color_to_rgb(unsigned int color)
 	int green;
 	int blue;
 
-	blue = color % 256;
-	green = ((color - blue) / 256) % 256;
-	red = ((color - blue) / (256 * 256)) - green / 256;
-	return (new_vector(red, green, blue));
+	blue = (int)(color % 256);
+	green = (int)(((color - blue) / 256) % 256);
+	red = (int)(((color - blue) / (256 * 256)) - green / 256);
+	return (new_vector((float)red, (float)green, (float)blue));
 }
 
 
@@ -29,8 +29,8 @@ int			rgb_to_color(s_vector *rgb)
 {
 	int color;
 
-	color = rgb->x;
-	color = (color << 8) + rgb->y;
-	color = (color << 8) + rgb->z;
+	color = (int)(((int)rgb->x));
+	color = (int)((color << 8) + (int)rgb->y);
+	color = (int)((color << 8) + (int)rgb->z);
 	return (color);
 }
